requirements.txt
beautifulsoup4==4.4.0
billiard==3.3.0.20
chardet==2.3.0
Coffin==0.3.8
Django==1.10
django-appconf==1.0.1
django-avatar==2.1.1
django-compressor==1.2
django-countries==3.3
django-haystack==2.6.0
django-keyedcache==1.5.1
django-kombu==0.9.4
django-picklefield==0.3.0
django-recaptcha-works==0.3.4
django-robots==1.1
django-transaction-signals==1.0.0
docutils==0.12
elasticsearch==2.0.0
funcsigs==0.4
html5lib==0.90
httplib2==0.9.1
Jinja2==2.8
kombu==3.0.26
lamson==1.3.4
lockfile==0.10.2
Logbook==0.7.0
longerusername==0.4
Markdown==2.4.1
markdown2==2.3.0
MarkupSafe==0.23
mock==1.3.0
nose==1.3.7
pbr==1.8.0
Pillow==7.0.0
Pygments==1.6
PyJWT==1.4.0
pystache==0.3.1
python-daemon==2.0.6
python-modargs==1.7
pytz==2013b0
redis==2.10.1
requests==2.3.0
sanction==0.3.1
simplejson==3.6.3
six==1.7.3
South==1.0
Unidecode==0.4.18
urllib3==1.9
verlib==0.1
wheel==0.24.0
MarkdownSuperscript==1.0.1
MarkdownSubscript==1.0.1
nltk==3.4.5
scikit-learn==0.20.4
